% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{B0}
\alias{B0}
\title{DATASET_TITLE}
\format{
A data frame with 2 rows and 2 variables:
\describe{
\item{\code{V1}}{double COLUMN_DESCRIPTION}
\item{\code{V2}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
B0
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
