% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterIndividual.R
\name{clusterIndividual}
\alias{clusterIndividual}
\title{Title}
\usage{
clusterIndividual(X, N0, m0, a0, c0, B0)
}
\arguments{
\item{X}{mama}

\item{N0}{lala}

\item{m0}{hha}

\item{a0}{kaka}

\item{c0}{uaua}

\item{B0}{yaya}
}
\value{
hjeke
}
\description{
Title
}
