% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension.R
\name{reduceDimension}
\alias{reduceDimension}
\title{Title}
\usage{
reduceDimension(D.sub, method, nrProjections)
}
\arguments{
\item{D.sub}{preprocessed matrix with top variable genes}

\item{nrProjections}{number of projections to be made by dimension reduction method}

\item{dim.method}{dimension reduction method among tSNE and UMAP}
}
\value{

}
\description{
Title
}
