% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X}
\alias{X}
\title{DATASET_TITLE}
\format{
A data frame with 1000 rows and 2 variables:
\describe{
\item{\code{X1}}{double COLUMN_DESCRIPTION}
\item{\code{X2}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
X
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
